Fix bugs, small improvements.
Now the modal window does not require the fsnavmenu selector and can work without it.

Demo styles have been moved to a separate file.
Added hidden h1.
Fixed a bug due to which fsnavmenu closed the header.
Fixed a bug due to which the minimum size was ignored in the adaptive font mixin. (How did I not notice this?)