New code-style for stylus, improvements to the demo page.
The use of the new code style is due to the reluctance to manually adjust the tabulation for styles.
(I'm really tired of this)
In this code style, when using the formatter (Manta's Stylus Supremacy), there are no problems with tabulation.
The following formatting parameters were used:
Insert Braces = true,
Insert Semicolons = true.