Big improvements to the submenu module.
Now the submenu can be opened either by clicking or hovering.

To reduce the amount of code (and selectors) for creating submenu buttons, they are now created in groups.

Added the ability to disable submenus enabled by click(!) via Esc.

Fixed a console output error when pressing Esc when the modal window is not open.

Fixed a bug that caused the minimum font size in the adapt-fsize mixin to not work.