Improvements in modules swipeMenu and sidebar.
Now the element of the touchArea does not need to be made large, the swipe is just beginning with it.
Swipe continues throughout the window until it is stopped or the menu switches.

After the menu has been activated, the touchArea can be changed because it is assigned the "active" class.

A bug has been fixed due to which the vertical swipe did not work.

And other minor changes.