Improvements to the swipe module.
Now you need to specify the width at which (and less than which) the swipe will work.
Thanks to this, an error was fixed, due to which the swipe did not work when resizing the window, but it worked after a reboot.
And yes, now the transform property can be omitted for the screen width above the one on which the swipe should work.