Applied esbuild to compile scripts.
This increased the compilation JS files speed by about 2 times.

Also now, in development mode, when compiling Sass files, css without the .min postfix is not created. (don't worry, they are created in production mode)
This was done because the files with postfix .min was no different from files without it in development mode.
Theoretically, this should increase the performance of the CSS task.