Improvements to the module animateByScroll.
Improved performance through the use of Intersection Observer and other minor improvements.
Now you can set the animateStartCoeff value not only for the beginning of the block, but also for its end.
That is, now you can configure the beginning of the animation when scrolling a block not only from its top, but also from its bottom.