Improved the experience of configuring pages.
Work has been carried out to improve the speed and convenience of setting up the page.
The bottom line is less code and a lot of settings right on the page.

Now you can quickly:
Connect script files (currently only working with imported ones).
Config the lang attribute.
Enter the name of a font to be pre-connected.