Separate loading general and specific styles.
Now styles for header, footer and general are loaded not together with the style file for a specific page, but separately, before loading styles for a specific page.

This was done so that there would be no need to connect a style file to a new page for index.html , where there are not only common, but also unnecessary styles.

TODO: Do same for styles and module scripts, like form and spoiler.