Optimized the process of adding scripts and styles for pages.
Now they can (and, in the case of styles, need to) be connected as strings in the EJS template array.
Otherwise, you can still connect scripts at the end of the file, after the body.

EJS templates for connecting modules and configuring common page attributes have been combined.

Also, the excess was removed from the development of the stepByStepBlock script.