[modalWindow] Improvements and correction of a critical error.
Fixed an error due to which the modal window could not be closed by clicking on the corresponding button.

Styles for modal windows have been moved to a separate file.

The example of a modal window has been slightly modified to better understand its structure.

Also now, when the window is closing, its content does not shift due to the sudden appearance of the page scroll.