From Stylus to postcss, part 1.
The stylus was abandoned, its functionality was replaced based on postcss plugins.
Stylus limited me in many ways, for example, in using new things in css (like media range syntax).

The task for working with css has been redesigned, now the configuration for postcss is not specified in the package.json, and in the plugin arguments.