[fix] Fixes and improvements.
Now the task for fonts should not slow down every time it starts.

Fixed an error due to which font names were incorrectly read in the automatic font-rule writing function.

Added stepByStep item to layout.html to control the stepByStepBlock connection.
In the same place, the modalWindow item was replaced with dialogs.

Fixed an error that caused the autoScrollPadding module to create an error when loading a page without a hash.