[swipe] Fixes and improvements.
Will I ever end working on it?

Optimizations.

Now in the swipe process, cursor becomes grabbing for the entire window and not just for swipeArea.

Fixed a bug that caused the maxWorkWidth parameter to not work.

Fixed a bug that could cause the swipe not to start.

Added automatic positioning of the element in the initial and final states depending on the ChangePlane parameter (will be improved and expanded).