[tab] Big improvements.
The structure of tabs has been changed. Now you should specify the aria-controls attribute for the buttons, where the ID of the tab to open is specified.

You can get acquainted with the new template (and test it) by running the project, see the file index.html.

Now the necessary role (tab, tablist, tabpanel) attributes are specified automatically.