[stepByStep] Fixes and improvements.
Fixed a bug that caused the block to be built incorrectly.

Added the ability to switch completed steps by clicking on the status buttons.

Added the ability to change the transition function between steps.

Added the ability to change the orientation of the movement of steps in the block.

Added support for preferences-reduced-motion.

Now, provided that the first or last step is the current one, the buttons for switching to the previous or next step are turned off, respectively.

Added support for aria attributes.

Added the ability to change the step that will become active by clicking on the status button. To do this, add the data-controls=block_number attribute.

Added class assignment for completed and active steps and status blocks.