[dialogs] Fixes and improvements.
Added README.md

The API file has been deleted, now to switch closure by backdrop, you need to specify the data attribute for the dialog, see the README.

Added support for aria-controls.

Fixed a bug where the aria-expanded attribute could be assigned incorrectly.

Fixed a bug that caused modals to open as just window.

Fixed a bug where closing on Esc was not handled.

The ability to switch the dialog by hovering the mouse and turning it off by moving the cursor outside the boundaries has been removed. This practice is incorrect.